seperated the tests into their corresponding class files
